% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argument.R
\name{arg}
\alias{arg}
\title{Deliver arguments of command}
\usage{
arg(tag, value, indicator = FALSE, lgl2int = FALSE, format = "\%s", sep = " ")
}
\arguments{
\item{tag}{A string specifying argument tag, like "-i", "-o".}

\item{value}{Value passed to the argument.}

\item{indicator}{A logical value specifying whether value should be an
indicator of tag. If TRUE, logical value will explain the set or unset of
tag.}

\item{lgl2int}{A logical value indicates whether transfrom value \code{TRUE} to 1
or \code{FALSE} to 0. If \code{TRUE}, format will always be set to "\%d".}

\item{format}{The format of the value, details see \link{sprintf}.}

\item{sep}{A character string used to separate "tag" and "value", usually " "
or "=".}
}
\value{
A string.
}
\description{
Deliver arguments of command
}
