% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biosys-package.R
\docType{package}
\name{biosys-package}
\alias{biosys}
\alias{biosys-package}
\title{biosys: System Command Function for Bioinformatics}
\description{
System Command function for Bioinformatics.
}
\author{
\strong{Maintainer}: Yun Peng \email{yunyunp96@163.com} (\href{https://orcid.org/0000-0003-2801-3332}{ORCID})

}
\keyword{internal}
