% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exec-cellranger.R
\name{cellranger}
\alias{cellranger}
\title{Run cellranger}
\usage{
cellranger(
  subcmd,
  ...,
  envpath = NULL,
  envvar = NULL,
  abort = TRUE,
  stdout = TRUE,
  stderr = TRUE,
  stdin = "",
  wait = TRUE,
  timeout = 0L,
  verbose = TRUE,
  help = FALSE,
  cellranger = NULL
)
}
\arguments{
\item{subcmd}{Sub-Command of cellranger.}

\item{...}{Additional arguments passed to \verb{cellranger subcmd} command.}

\item{envpath}{A character define the environment variables \code{PATH} to be
added before running command.}

\item{envvar}{A named atomic vector define running environment variables of
the command, all environment variables will be replaced before running
command. Use \code{NA} to remove an environment variable.}

\item{abort}{A bool indicates whether to report error if command return
non-zero status.}

\item{stdout, stderr}{Where output to ‘stdout’ or ‘stderr’ should be sent.
Possible values are:
\itemize{
\item \code{TRUE}: print child output in R console
\item \code{FALSE}: suppress output stream
\item \strong{string}: name or path of file to redirect output
}}

\item{stdin}{should input be diverted?  \code{""} means the default,
    alternatively a character string naming a file.  Ignored
    if \code{input} is supplied.}

\item{wait}{A bool (not NA) indicating whether the R interpreter should wait
for the command to finish, or run it asynchronously. This will be ignored
(and the interpreter will always wait) if \code{abort = TRUE}.}

\item{timeout}{timeout in seconds, ignored if 0.  This is a limit for the
    elapsed time running \code{command} in a separate process.   Fractions
    of seconds are ignored.}

\item{verbose}{A logical value indicates whether to print running command
message.}

\item{help}{If \code{TRUE}, will print the help document for this command.}

\item{cellranger}{A string of path to \code{cellranger} command.}
}
\description{
Run cellranger
}
