% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd-.R
\name{cmd_help}
\alias{cmd_help}
\alias{cmd_run}
\title{Execute command}
\usage{
cmd_help(command, verbose = TRUE)

cmd_run(
  command,
  stdout = TRUE,
  stderr = TRUE,
  stdin = NULL,
  timeout = NULL,
  verbose = TRUE
)
}
\arguments{
\item{command}{A \code{command} object.}

\item{verbose}{A bool indicates whether to print running command message.}

\item{stdout, stderr}{Where output to 'stdout' or 'stderr' should be
sent. Possible values are:
\itemize{
\item \code{TRUE}: print child output in R console
\item \code{FALSE}: suppress output stream
\item \strong{string}: name or path of file to redirect output
}}

\item{stdin}{should input be diverted? \code{NULL} means the default,
alternatively a character string naming a file. Ignored if input is
supplied.}

\item{timeout}{Timeout in seconds, ignored if \code{NULL}. This is a limit for
the elapsed time running command in a separate process. Fractions of
seconds are ignored.}
}
\value{
\itemize{
\item \code{cmd_help}: the input \code{command}.
}

\itemize{
\item \code{cmd_run}: Exit status.
}
}
\description{
\itemize{
\item \code{cmd_help}: Print the help document for this command.
\item \code{cmd_run}: Run the command.
}
}
\seealso{
\code{\link[=cmd_wd]{cmd_wd()}}/\code{\link[=cmd_envvar]{cmd_envvar()}}/\code{\link[=cmd_envpath]{cmd_envpath()}}
}
