% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd-.R
\name{cmd_run}
\alias{cmd_run}
\alias{cmd_help}
\title{Execute command}
\usage{
cmd_run(
  command,
  stdout = NULL,
  stderr = NULL,
  stdin = NULL,
  timeout = NULL,
  verbose = TRUE
)

cmd_help(command, stdout = NULL, stderr = NULL, verbose = TRUE)
}
\arguments{
\item{command}{A \code{command} object.}

\item{stdout, stderr}{Where output to 'stdout' or 'stderr' should be
sent. Possible values are:
\itemize{
\item \code{NULL}: print child output in R console
\item \code{TRUE}: capture the output in a character vector
\item \code{FALSE}: suppress output stream
\item \strong{string}: name or path of file to redirect output
}}

\item{stdin}{should input be diverted? \code{NULL} means the default,
alternatively a character string naming a file. Ignored if input is
supplied.}

\item{timeout}{Timeout in seconds, ignored if \code{NULL}. This is a limit for
the elapsed time running command in a separate process. Fractions of
seconds are ignored.}

\item{verbose}{A single boolean value indicates whether to print running
command message.}
}
\value{
\itemize{
\item \code{cmd_run}: Exit status or the captured output when \code{stdout}/\code{stderr} is
\code{TRUE}.
}

\itemize{
\item \code{cmd_help}: the input \code{command} or the captured output when
\code{stdout}/\code{stderr} is \code{TRUE}.
}
}
\description{
\itemize{
\item \code{cmd_run}: Run the command.
\item \code{cmd_help}: Print the help document for this command.
}
}
\seealso{
\code{\link[=cmd_wd]{cmd_wd()}}/\code{\link[=cmd_envvar]{cmd_envvar()}}/\code{\link[=cmd_envpath]{cmd_envpath()}}
}
