% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exec-compress.R
\name{compress}
\alias{compress}
\alias{decompress}
\title{Python is a programming language that lets you work quickly and integrate
systems more effectively.}
\usage{
compress(
  mode,
  file,
  ...,
  ofile = NULL,
  odir = getwd(),
  keep = TRUE,
  override = FALSE,
  envpath = NULL,
  envvar = NULL,
  help = FALSE,
  stdout = TRUE,
  stderr = TRUE,
  stdin = "",
  wait = TRUE,
  timeout = 0L,
  abort = TRUE,
  verbose = TRUE,
  cmd = NULL
)

decompress(mode, file, ..., ofile = NULL)
}
\arguments{
\item{mode}{One of \code{gz}, \code{gzip}, \code{pigz}, \code{bzip2}, or \code{xz}.}

\item{file}{A string of input file path.}

\item{...}{<\link[rlang:dyn-dots]{dynamic dots}> Additional arguments passed to \code{compress} command, empty arguments are automatically trimed. Details see: \code{compress(cmd, help = TRUE)}.}

\item{ofile}{A string of path to the output file.}

\item{odir}{A string of path to the output directory.}

\item{keep}{A bool, indicates whether keep original file.}

\item{override}{A bool, indicates whether to override ouput file if it
exists.}

\item{envpath}{A character define the environment variables \code{PATH} to be
\code{added} before running command. See \link[withr:with_path]{withr::with_path}.}

\item{envvar}{A named atomic vector define running environment variables of
the command, all environment variables will be \code{replaced} before running
command. See \link[withr:with_envvar]{withr::with_envvar}.}

\item{help}{If \code{TRUE}, will print the help document for this command.}

\item{stdout, stderr}{Where output to ‘stdout’ or ‘stderr’ should be sent.
Possible values are:
\itemize{
\item \code{TRUE}: print child output in R console
\item \code{FALSE}: suppress output stream
\item \strong{string}: name or path of file to redirect output
}}

\item{stdin}{should input be diverted?  \code{""} means the default,
    alternatively a character string naming a file.  Ignored
    if \code{input} is supplied.}

\item{wait}{A bool (not NA) indicating whether the R interpreter should wait
for the command to finish, or run it asynchronously  Only used when
\code{help=FALSE}. When wait is not \code{TRUE}, it can also be a string to name the
process, the asynchronous process can be managed with \link{process} tools.}

\item{timeout}{timeout in seconds, ignored if 0.  This is a limit for the
    elapsed time running \code{command} in a separate process.   Fractions
    of seconds are ignored.}

\item{abort}{A bool indicates whether to report error if command return
non-zero status.}

\item{verbose}{A bool indicates whether to print running command message.}

\item{cmd}{Command path to be invoked, as a character string.}
}
\description{
Python is a programming language that lets you work quickly and integrate
systems more effectively.
}
\seealso{
\itemize{
\item \url{https://www.gnu.org/software/gzip/}
\item \url{https://zlib.net/pigz/}
\item \url{https://sourceware.org/bzip2/}
\item \url{https://tukaani.org/xz/}
}
}
