% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exec.R
\name{exec}
\alias{exec}
\title{Invoke a System Command}
\usage{
exec(
  cmd,
  ...,
  envpath = NULL,
  envvar = NULL,
  opath = NULL,
  stdout = TRUE,
  stderr = TRUE,
  stdin = "",
  wait = TRUE,
  timeout = 0L,
  abort = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{cmd}{Command to be invoked, as a character string.}

\item{...}{<\link[rlang:dyn-dots]{dynamic dots}> Arguments passed to \code{cmd}.}

\item{envpath}{A character define the environment variables \code{PATH} to be
added before running command.}

\item{envvar}{A named atomic vector define running environment variables of
the command, all environment variables will be replaced before running
command. Use \code{NA} to remove an environment variable.}

\item{opath}{Specifying the output file or directory to be removed if
command running error. Note: all files in the directory will be removed, you
must use this argument carefully.}

\item{stdout, stderr}{Where output to ‘stdout’ or ‘stderr’ should be sent.
Possible values are:
\itemize{
\item \code{TRUE}: print child output in R console
\item \code{FALSE}: suppress output stream
\item \strong{string}: name or path of file to redirect output
}}

\item{stdin}{should input be diverted?  \code{""} means the default,
    alternatively a character string naming a file.  Ignored
    if \code{input} is supplied.}

\item{wait}{A bool (not NA) indicating whether the R interpreter should wait
for the command to finish, or run it asynchronously  Only used when
\code{help=FALSE}. The asynchronous process can be managed with \link{process} tools.}

\item{timeout}{timeout in seconds, ignored if 0.  This is a limit for the
    elapsed time running \code{command} in a separate process.   Fractions
    of seconds are ignored.}

\item{abort}{A bool indicates whether to report error if command return
non-zero status.}

\item{verbose}{A bool indicates whether to print running command message.}
}
\value{
\itemize{
\item if \code{wait=FALSE}, the process ID.
\item if \code{abort=TRUE} and \code{wait=TRUE}, zero if command success, otherwise, abort
error.
\item if \code{abort=FALSE} and \code{wait=TRUE}, exit status returned by the command.
}
}
\description{
Invoke a System Command
}
