% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exec-fastq_pair.R
\name{fastq_pair}
\alias{fastq_pair}
\alias{fastq_read_pair}
\title{FASTQ PAIR}
\usage{
fastq_pair(
  fq1,
  fq2,
  ...,
  hash_table_size = NULL,
  keep_decompressed = FALSE,
  keep_unpaired = TRUE,
  compress = TRUE,
  odir = getwd(),
  envpath = NULL,
  envvar = NULL,
  help = FALSE,
  stdout = TRUE,
  stderr = TRUE,
  stdin = "",
  wait = TRUE,
  timeout = 0L,
  abort = TRUE,
  verbose = TRUE,
  fastq_pair = NULL
)

fastq_read_pair(fastq_files)
}
\arguments{
\item{fq1, fq2}{A string of fastq file path.}

\item{...}{Additional arguments passed to \code{fastq_pair} command. Details see: \code{fastq_pair(help = TRUE)}.}

\item{hash_table_size}{Size of hash table to use.}

\item{keep_decompressed}{A bool, If \code{fq1} or \code{fq2} are compressed,
\code{fastq_pair} function will automatically decompress them, this argument
controls whether we should remove the decompressed temporary files.}

\item{keep_unpaired}{A bool indicates whether to keep the unpaired sequence.}

\item{compress}{A bool, whether compress final results.}

\item{odir}{A string of path to the output directory.}

\item{envpath}{A character define the environment variables \code{PATH} to be
added before running command.}

\item{envvar}{A named atomic vector define running environment variables of
the command, all environment variables will be replaced before running
command. Use \code{NA} to remove an environment variable.}

\item{help}{If \code{TRUE}, will print the help document for this command.}

\item{stdout, stderr}{Where output to ‘stdout’ or ‘stderr’ should be sent.
Possible values are:
\itemize{
\item \code{TRUE}: print child output in R console
\item \code{FALSE}: suppress output stream
\item \strong{string}: name or path of file to redirect output
}}

\item{stdin}{should input be diverted?  \code{""} means the default,
    alternatively a character string naming a file.  Ignored
    if \code{input} is supplied.}

\item{wait}{A bool (not NA) indicating whether the R interpreter should wait
for the command to finish, or run it asynchronously. This will be ignored
(and the interpreter will always wait) if \code{abort = TRUE}.}

\item{timeout}{timeout in seconds, ignored if 0.  This is a limit for the
    elapsed time running \code{command} in a separate process.   Fractions
    of seconds are ignored.}

\item{abort}{A bool indicates whether to report error if command return
non-zero status.}

\item{verbose}{A bool indicates whether to print running command message.}

\item{fastq_pair}{A string of path to \code{fastq_pair} command.}

\item{fastq_files}{A character of the fastq file paths.}
}
\description{
Rewrite paired end fastq files to make sure that all reads have a mate and to
separate out singletons.

Usually when you get paired end read files you have two files with a /1
sequence in one and a /2 sequence in the other (or a /f and /r or just two
reads with the same ID). However, often when working with files from a third
party source (e.g. the SRA) there are different numbers of reads in each file
(because some reads fail QC). Spades, bowtie2 and other tools break because
they demand paired end files have the same number of reads.
}
\seealso{
\url{https://github.com/linsalrob/fastq-pair}
}
