% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd-gistic2.R
\name{gistic2}
\alias{gistic2}
\title{Run GISTIC2}
\usage{
gistic2(
  seg,
  refgene,
  ...,
  odir = getwd(),
  envpath = NULL,
  envvar = NULL,
  abort = TRUE,
  stdout = TRUE,
  stderr = TRUE,
  stdin = "",
  wait = TRUE,
  timeout = 0L,
  verbose = TRUE,
  help = FALSE,
  gistic2 = NULL
)
}
\arguments{
\item{seg}{A data.frame of segmented data.}

\item{refgene}{Path to reference genome data input file (REQUIRED, see below
for file description).}

\item{...}{Other arguments passed to \code{gistic2}. Details see:
\url{https://broadinstitute.github.io/gistic2/}.}

\item{odir}{Output directory.}

\item{envpath}{A character define the environment variables \code{PATH} to be
added before running command.}

\item{envvar}{A named atomic vector define running environment variables of
the command.}

\item{abort}{A bool indicates whether to report error if command return
non-zero status.}

\item{stdout, stderr}{Where output to ‘stdout’ or ‘stderr’ should be sent.
Possible values are:
\itemize{
\item \code{TRUE}: print child output in R console
\item \code{FALSE}: suppress output stream
\item \strong{string}: name or path of file to redirect output
}}

\item{stdin}{should input be diverted?  \code{""} means the default,
    alternatively a character string naming a file.  Ignored
    if \code{input} is supplied.}

\item{wait}{A bool (not NA) indicating whether the R interpreter should wait
for the command to finish, or run it asynchronously. This will be ignored
(and the interpreter will always wait) if \code{abort = TRUE}.}

\item{timeout}{timeout in seconds, ignored if 0.  This is a limit for the
    elapsed time running \code{command} in a separate process.   Fractions
    of seconds are ignored.}

\item{verbose}{A logical value indicates whether to print running command
message.}

\item{help}{If \code{TRUE}, will print the help document for this command.}

\item{gistic2}{A string of path to \code{gistic2} command.}
}
\description{
The GISTIC module identifies regions of the genome that are
significantly amplified or deleted across a set of samples. Each aberration
is assigned a G-score that considers the amplitude of the aberration as well
as the frequency of its occurrence across samples. False Discovery Rate
q-values are then calculated for the aberrant regions, and regions with
q-values below a user-defined threshold are considered significant. For each
significant region, a "peak region" is identified, which is the part of the
aberrant region with greatest amplitude and frequency of alteration. In
addition, a "wide peak" is determined using a leave-one-out algorithm to
allow for errors in the boundaries in a single sample. The "wide peak"
boundaries are more robust for identifying the most likely gene targets in
the region. Each significantly aberrant region is also tested to determine
whether it results primarily from broad events (longer than half a chromosome
arm), focal events, or significant levels of both. The GISTIC module reports
the genomic locations and calculated q-values for the aberrant regions. It
identifies the samples that exhibit each significant amplification or
deletion, and it lists genes found in each "wide peak" region.
}
\seealso{
\url{https://broadinstitute.github.io/gistic2/}
}
