% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exec-kraken2.R
\name{kraken2}
\alias{kraken2}
\title{Running Kraken2}
\usage{
kraken2(
  fq1,
  ...,
  fq2 = NULL,
  classified_out = NULL,
  envpath = NULL,
  envvar = NULL,
  abort = TRUE,
  stdout = TRUE,
  stderr = TRUE,
  stdin = "",
  wait = TRUE,
  timeout = 0L,
  verbose = TRUE,
  help = FALSE,
  kraken2 = NULL
)
}
\arguments{
\item{fq1, fq2}{Path to fastq 1 file.}

\item{...}{Additional arguments passed to \code{kraken2} command. Details see: \code{kraken2(help = TRUE)}}

\item{classified_out}{A string of path to save classified sequences.}

\item{envpath}{A character define the environment variables \code{PATH} to be
added before running command.}

\item{envvar}{A named atomic vector define running environment variables of
the command, all environment variables will be replaced before running
command. Use \code{NA} to remove an environment variable.}

\item{abort}{A bool indicates whether to report error if command return
non-zero status.}

\item{stdout, stderr}{Where output to ‘stdout’ or ‘stderr’ should be sent.
Possible values are:
\itemize{
\item \code{TRUE}: print child output in R console
\item \code{FALSE}: suppress output stream
\item \strong{string}: name or path of file to redirect output
}}

\item{stdin}{should input be diverted?  \code{""} means the default,
    alternatively a character string naming a file.  Ignored
    if \code{input} is supplied.}

\item{wait}{A bool (not NA) indicating whether the R interpreter should wait
for the command to finish, or run it asynchronously. This will be ignored
(and the interpreter will always wait) if \code{abort = TRUE}.}

\item{timeout}{timeout in seconds, ignored if 0.  This is a limit for the
    elapsed time running \code{command} in a separate process.   Fractions
    of seconds are ignored.}

\item{verbose}{A logical value indicates whether to print running command
message.}

\item{help}{If \code{TRUE}, will print the help document for this command.}

\item{kraken2}{A string of path to \code{kraken2} command.}
}
\description{
Metagenomic classification of paired-end reads from single-cell RNA
sequencing fastq files can be performed using any k-mer based mapper that
identifies a taxonomic ID for each k-mer and read. However, \code{SAHMI} is
optimized to run with \code{Kraken2Uniq}, which finds exact matches of candidate
35-mer genomic substrings to the lowest common ancestor of genomes in a
reference metagenomic database. It is essential that all realistically
possible genomes are included as mapping references at this stage (e.g. host,
known vectors, etc.), or that host mappable reads are excluded. The required
outputs from this step are: a Kraken summary report with sample level
metagenomic counts, a Kraken output file with read and k-mer level taxonomic
classifications, an MPA-style report, and raw sequencing fastq files with
taxonomic classification for each read. Please see
\url{https://github.com/DerrickWood/kraken2/blob/master/docs/MANUAL.markdown} for
more details on installation and usage of Kraken2/KrakenUniq.
}
\seealso{
\url{https://github.com/DerrickWood/kraken2/wiki/Manual}
}
