% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exec-kraken2-tools.R
\name{kraken_tools}
\alias{kraken_tools}
\title{KrakenTools is a suite of scripts to be used alongside the Kraken,
KrakenUniq, Kraken 2, or Bracken programs.}
\usage{
kraken_tools(
  script,
  ...,
  pythonpath = NULL,
  envpath = NULL,
  envvar = NULL,
  help = FALSE,
  stdout = TRUE,
  stderr = TRUE,
  stdin = "",
  wait = TRUE,
  timeout = 0L,
  abort = TRUE,
  verbose = TRUE,
  python = NULL
)
}
\arguments{
\item{script}{Name of the kraken2 script. One of
\code{combine_kreports}, \code{combine_mpa}, \code{extract_kraken_reads}, \code{filter_bracken_out}, \code{fix_unmapped}, \code{kreport2krona}, \code{kreport2mpa}, \code{make_kreport}, and \code{make_ktaxonomy}}

\item{...}{<\link[rlang:dyn-dots]{dynamic dots}> Additional arguments passed to \code{script} command, empty arguments are automatically trimed. Details see: \code{kraken_tools(script, help = TRUE)}.}

\item{pythonpath}{A character define the environment variables \code{PYTHONPATH}
to be \code{added} before running command.}

\item{envpath}{A character define the environment variables \code{PATH} to be
\code{added} before running command. See \link[withr:with_path]{withr::with_path}.}

\item{envvar}{A named atomic vector define running environment variables of
the command, all environment variables will be \code{replaced} before running
command. See \link[withr:with_envvar]{withr::with_envvar}.}

\item{help}{If \code{TRUE}, will print the help document for this command.}

\item{stdout, stderr}{Where output to ‘stdout’ or ‘stderr’ should be sent.
Possible values are:
\itemize{
\item \code{TRUE}: print child output in R console
\item \code{FALSE}: suppress output stream
\item \strong{string}: name or path of file to redirect output
}}

\item{stdin}{should input be diverted?  \code{""} means the default,
    alternatively a character string naming a file.  Ignored
    if \code{input} is supplied.}

\item{wait}{A bool (not NA) indicating whether the R interpreter should wait
for the command to finish, or run it asynchronously  Only used when
\code{help=FALSE}. When wait is not \code{TRUE}, it can also be a string to name the
process, the asynchronous process can be managed with \link{process} tools.}

\item{timeout}{timeout in seconds, ignored if 0.  This is a limit for the
    elapsed time running \code{command} in a separate process.   Fractions
    of seconds are ignored.}

\item{abort}{A bool indicates whether to report error if command return
non-zero status.}

\item{verbose}{A bool indicates whether to print running command message.}

\item{python}{A string of path to \code{python} command.}
}
\description{
These scripts are designed to help Kraken users with downstream
analysis of Kraken results.
}
\seealso{
\url{https://github.com/jenniferlu717/KrakenTools}
}
