% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envvars.R
\name{parse_envvars}
\alias{parse_envvars}
\title{Parse an atomic character for system2 env argument}
\usage{
parse_envvars(env)
}
\arguments{
\item{env}{A named atomic character}
}
\value{
A string can be used in system2
}
\description{
Parse an atomic character for system2 env argument
}
