% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process}
\alias{process}
\alias{process_echo}
\alias{process_is_running}
\alias{process_is_finished}
\alias{process_kill}
\alias{process_value}
\title{Biosys asynchronous process management}
\usage{
process_echo()

process_is_running(x)

process_is_finished(x)

process_kill(x)

process_value(x, clean = TRUE)
}
\arguments{
\item{x}{Process names or an integer vector of process ID. For
\code{process_value}, only 1 process is supporeted.}

\item{clean}{A bool, indicates whether to remove collected value.}
}
\value{
\itemize{
\item \code{process_echo}: A logical value indicates if process were running.
}

\itemize{
\item \code{process_is_running}: A logical value indicates if process were running.
}

\itemize{
\item \code{process_is_finished}: A logical value indicates if process were finished.
}

\itemize{
\item \code{process_kill}: A logical value indicates if process were killed
successfully.
}

\itemize{
\item \code{process_value}: Value returned by process command.
}
}
\description{
\itemize{
\item \code{process_echo}: Show \code{Biosys} background process.
\item \code{process_is_running}: Check if the process is running.
\item \code{process_is_finished}: Check if the process is finished.
\item \code{process_kill}: Terminate the process.
\item \code{process_value}: Extract process value.
}
}
