% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exec-pyscenic.R
\name{pyscenic}
\alias{pyscenic}
\title{Run pyscenic}
\usage{
pyscenic(subcmd = NULL, ..., pyscenic = NULL)
}
\arguments{
\item{subcmd}{Sub-Command of pyscenic.}

\item{...}{<\link[rlang:dyn-dots]{dynamic dots}> Additional arguments passed to \verb{pyscenic subcmd} command, empty arguments are automatically trimed.}

\item{pyscenic}{A string of path to \code{pyscenic} command.}
}
\value{
A \link{Execute} object.
}
\description{
Run pyscenic
}
\references{
\url{https://github.com/aertslab/pySCENIC}
}
\seealso{
\code{\link[=run_pyscenic]{run_pyscenic()}}
}
