% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envvars.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{with_envvar}
\alias{with_path}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{withr}{\code{\link[withr]{with_envvar}}, \code{\link[withr]{with_path}}}
}}

