% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alleleCounter.R
\name{run_allele_counter}
\alias{run_allele_counter}
\title{Run alleleCount}
\usage{
run_allele_counter(
  hts_file,
  loci_file,
  output_file,
  ref_fasta = NULL,
  min_base_qual = 20L,
  min_map_qual = 35L,
  dense_snp = TRUE,
  ...,
  allelecounter_cmd = NULL,
  envpath = NULL,
  env = NULL,
  sys_args = list(),
  verbose = TRUE
)
}
\arguments{
\item{hts_file}{Path to sample HTS file.}

\item{loci_file}{Path to loci file.}

\item{output_file}{Path write output file.}

\item{ref_fasta}{Path to reference fasta index file.}

\item{min_base_qual}{Minimum base quality. Default: 20.}

\item{min_map_qual}{Minimum mapping quality. Default: 35.}

\item{dense_snp}{Improves performance where many positions are close
together.}

\item{...}{Other arguments passed to alleleCounter}

\item{allelecounter_cmd}{Path to alleleCounter command.}

\item{envpath}{A character define the environment PATH to add before running
command.}

\item{env}{A named atomic vector define running environment of the command.}

\item{sys_args}{A list of arguments passed to \link{system2}.}

\item{verbose}{A logical value indicates output running command message.}
}
\description{
The alleleCount package primarily exists to prevent code duplication between
some other projects, specifically AscatNGS and Battenberg.
}
\seealso{
https://github.com/cancerit/alleleCount
}
