% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellranger.R
\name{run_cellranger}
\alias{run_cellranger}
\title{Run cellranger}
\usage{
run_cellranger(
  subcmd,
  args = character(),
  cellranger_cmd = NULL,
  env_path = NULL,
  env = character(),
  sys_args = list(),
  verbose = TRUE
)
}
\arguments{
\item{subcmd}{Sub-Command of cellranger.}

\item{args}{Arguments passed to cellranger sub-command.}

\item{cellranger_cmd}{Path to cellranger command.}

\item{env_path}{An atomic vector define the \code{PATH} to be added into the
environment variable \code{PATH}.}

\item{env}{A named character define the to environment variables when running
\code{cmd}.}

\item{sys_args}{A list of arguments passed to \link{system2}.}

\item{verbose}{A logical value indicates output running command message.}
}
\description{
Run cellranger
}
