% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conipher_cluster.R
\name{run_conipher_cluster}
\alias{run_conipher_cluster}
\title{Running CONIPHER Clustering}
\usage{
run_conipher_cluster(
  mut_tsv,
  seg_tsv,
  case_id,
  output_dir = getwd(),
  subclonal_copy_correction = TRUE,
  only_truncal_subclonal_copy_correction = TRUE,
  run_pyclone = TRUE,
  pyclone_cmd = NULL,
  pyclone_yaml = NULL,
  min_cluster_size = 5L,
  multiple_test_correction = TRUE,
  clean_clusters = TRUE,
  clonal_cutOff = 0.9,
  propClonal_threshold = 0.25,
  fix_absentCCFs = TRUE,
  driver_filter = "1A,1,2A",
  burn_in = 1000L,
  seed = 1024L,
  cores = 1L,
  envpath = NULL,
  env = NULL,
  sys_args = list()
)
}
\arguments{
\item{mut_tsv}{A mutation table in long format (mutations x tumour regions).}

\item{seg_tsv}{A copy number segment .tsv used for plotting only.}

\item{case_id}{A string specifying the tumour case ID.}

\item{output_dir}{Path specifying the results output directory.}

\item{subclonal_copy_correction}{should subclonal copy number correction be
used.}

\item{only_truncal_subclonal_copy_correction}{should only truncal subclonal
copy number correction be used.}

\item{run_pyclone}{A logical, whether to run pyclone.}

\item{pyclone_cmd}{The path of pyclone.}

\item{pyclone_yaml}{A template yaml file for pyclone.}

\item{min_cluster_size}{minimum number of mutations in a cluster to be
considered.}

\item{multiple_test_correction}{should multiple testing correction be applied
for the copy number correcting mutations.}

\item{clean_clusters}{should clusters be cleaned and merged?}

\item{clonal_cutOff}{lower threshold CCF to be considered clonal.}

\item{propClonal_threshold}{proportion of cluster that needs to be considered
clonal to merge.}

\item{fix_absentCCFs}{should CCF of absent mutations be set to zero?}

\item{driver_filter}{what filter to use for drivers.}

\item{burn_in}{burn-in for DP clustering.}

\item{seed}{Random seed for pyclone.}

\item{cores}{Number of cores allocated to run script in parallel.}

\item{envpath}{A character define the environment PATH to add before running
command.}

\item{env}{A named atomic vector define running environment of the command.}

\item{sys_args}{A list of arguments passed to \link{system2}.}
}
\description{
Running CONIPHER Clustering
}
\seealso{
https://github.com/McGranahanLab/CONIPHER-wrapper
}
