% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/command-trust4.R
\name{run_trust4}
\alias{run_trust4}
\alias{trust4_imgt_annot}
\alias{trust4_gene_names}
\title{TRUST4: immune repertoire reconstruction from bulk and single-cell RNA-seq
data}
\usage{
run_trust4(
  file1,
  ref_coordinate,
  ...,
  file2 = NULL,
  type = NULL,
  ref_annot = NULL,
  ofile = NULL,
  odir = getwd(),
  trust4_cmd = NULL,
  envpath = NULL,
  env = NULL,
  abort = FALSE,
  sys_args = list(),
  verbose = TRUE
)

trust4_imgt_annot(
  species = "Homo_sapien",
  ofile = "IMGT+C.fa",
  odir = getwd(),
  perl_cmd = NULL,
  envpath = NULL,
  env = NULL,
  abort = FALSE,
  sys_args = list(),
  verbose = TRUE
)

trust4_gene_names(
  imgt_annot,
  ofile = "bcr_tcr_gene_name.txt",
  odir = getwd(),
  grep_cmd = NULL,
  envpath = NULL,
  env = NULL,
  abort = FALSE,
  sys_args = list(),
  verbose = TRUE
)
}
\arguments{
\item{file1}{Path to bam file or fastq file.}

\item{ref_coordinate}{Path to the fasta file coordinate and sequence of
V/D/J/C genes.}

\item{...}{Other arguments passed to \code{run-trust4}. Details see:
\url{https://github.com/liulab-dfci/TRUST4}.}

\item{file2}{Path to the second paired-end read fastq file, only used for
\code{type = "fastq"}.}

\item{type}{One of "bam" or "fastq".}

\item{ref_annot}{Path to detailed V/D/J/C gene reference file, such as from
IMGT database. (default: not used). (recommended).}

\item{ofile}{\itemize{
\item \code{run_trust4}: Prefix of output files. (default: inferred from file
prefix).
\item \code{trust4_imgt_annot}: Output file name.
\item \code{trust4_gene_names}: Output file name.
}}

\item{odir}{Output directory.}

\item{trust4_cmd}{Path to \code{run-trust4} command.}

\item{envpath}{A character define the environment PATH to add before running
command.}

\item{env}{A named atomic vector define running environment of the command.}

\item{abort}{A scalar logical indicates whether to report error if command
return non-zero status.}

\item{sys_args}{A list of arguments passed to \link{system2}.}

\item{verbose}{A logical value indicates output running command message.}

\item{species}{Species to extract IMGT annotation, details see
\url{https://www.imgt.org//download/V-QUEST/IMGT_V-QUEST_reference_directory/}.}

\item{perl_cmd}{Path to \code{perl} command.}

\item{imgt_annot}{Path of IMGT annotation file, created via
\code{trust4_imgt_annot}.}

\item{grep_cmd}{Path to \code{grep} command.}
}
\description{
TRUST4: immune repertoire reconstruction from bulk and single-cell RNA-seq
data
}
\seealso{
\url{https://github.com/liulab-dfci/TRUST4}
}
