% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd-seqkit.R
\name{seqkit}
\alias{seqkit}
\title{Run seqkit}
\usage{
seqkit(subcmd = NULL, ..., seqkit = NULL)
}
\arguments{
\item{subcmd}{Sub-Command of seqkit.}

\item{...}{<\link[rlang:dyn-dots]{dynamic dots}> Additional arguments passed to \verb{seqkit subcmd} command, empty arguments are automatically trimed. Details see: \verb{cmd_help(seqkit subcmd())}.}

\item{seqkit}{A string of path to \code{seqkit} command.}
}
\value{
A \code{command} object.
}
\description{
Run seqkit
}
\seealso{
\url{https://bioinf.shenwei.me/seqkit/}
}
