% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envvars.R
\name{with_envvars}
\alias{with_envvars}
\title{Temporarily change system environment variables.}
\usage{
with_envvars(env, call, action = "replace", sep = .Platform$path.sep)
}
\arguments{
\item{env}{A named character}

\item{call}{Code to execute in the temporary environment}

\item{action}{should \code{env} values "replace", "prefix" or "suffix" existing
variables with the same name.}

\item{sep}{A string separates the elements in the environment variable.}
}
\description{
Temporarily change system environment variables.
}
