% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exec-cellranger.R
\name{cellranger}
\alias{cellranger}
\title{Run cellranger}
\usage{
cellranger(subcmd = NULL, ..., cellranger = NULL)
}
\arguments{
\item{subcmd}{Sub-Command of cellranger.}

\item{...}{<\link[rlang:dyn-dots]{dynamic dots}> Additional arguments passed to \verb{cellranger subcmd} command, empty arguments are automatically trimed.}

\item{cellranger}{A string of path to \code{cellranger} command.}
}
\description{
Run cellranger
}
