% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd-cellranger.R
\name{cellranger}
\alias{cellranger}
\title{Run cellranger}
\usage{
cellranger(subcmd = NULL, ..., cellranger = NULL)
}
\arguments{
\item{subcmd}{Sub-Command of cellranger.}

\item{...}{<\link[rlang:dyn-dots]{dynamic dots}> Additional arguments passed to \verb{cellranger subcmd} command, empty arguments are automatically trimed. Details see: \verb{cmd_help(cellranger subcmd())}.}

\item{cellranger}{A string of path to \code{cellranger} command.}
}
\value{
A \code{command} object.
}
\description{
Run cellranger
}
