% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exec-.R
\name{exec}
\alias{exec}
\title{Invoke a System Command}
\usage{
exec(cmd, ...)
}
\arguments{
\item{cmd}{Command to be invoked, as a character string.}

\item{...}{<\link[rlang:dyn-dots]{dynamic dots}> Additional arguments passed to \code{cmd} command, empty arguments are automatically trimed.}
}
\value{
A \link{Execute} object.
}
\description{
Invoke a System Command
}
