% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_arg.R
\name{handle_arg}
\alias{handle_arg}
\title{Deliver arguments of command}
\usage{
handle_arg(
  tag,
  value = TRUE,
  indicator = is_scalar_logical(value),
  lgl2int = FALSE,
  format = "\%s",
  sep = " "
)
}
\arguments{
\item{tag}{A string specifying argument tag, like "-i", "-o".}

\item{value}{Value passed to the argument.}

\item{indicator}{A logical value specifying whether value should be a
ndicator of tag. If TRUE, logical value will explain the set (TRUE or NULL)
or unset (FALSE) of tag.}

\item{lgl2int}{A logical value indicates whether transfrom value TRUE to 1 or
FALSE to 0. If \code{TRUE}, format will always be set to "\%d".}

\item{format}{The format of the value, details see \link{sprintf}.}

\item{sep}{A character string used to separate "tag" and "value", usually " "
or "=".}
}
\value{
A string or NULL which can be used directly by \link{run_command}.
}
\description{
Deliver arguments of command
}
