% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exec-kraken2.R
\name{kraken2}
\alias{kraken2}
\alias{kraken2_mpa}
\title{Running Kraken2}
\usage{
kraken2(
  fq1,
  ...,
  fq2 = NULL,
  ofile = "kraken_output.txt",
  report = "kraken_report.txt",
  classified_out = NULL,
  unclassified_out = NULL,
  odir = getwd(),
  envpath = NULL,
  envvar = NULL,
  abort = TRUE,
  stdout = TRUE,
  stderr = TRUE,
  stdin = "",
  wait = TRUE,
  timeout = 0L,
  verbose = TRUE,
  help = FALSE,
  kraken2 = NULL
)

kraken2_mpa(
  report,
  ...,
  ofile = NULL,
  odir = getwd(),
  pythonpath = NULL,
  envpath = NULL,
  envvar = NULL,
  abort = TRUE,
  stdout = TRUE,
  stderr = TRUE,
  stdin = "",
  wait = TRUE,
  timeout = 0L,
  verbose = TRUE,
  help = FALSE,
  python = NULL
)
}
\arguments{
\item{fq1, fq2}{A string of fastq file path.}

\item{...}{Additional arguments passed to \code{kraken2} command. Details see: \code{kraken2(help = TRUE)}}

\item{ofile}{A string of path to save kraken2 output.}

\item{report}{A string of path or the report file returned by \code{kraken2}.}

\item{classified_out}{A string of path to save classified sequences, which
should be a fastq file.}

\item{unclassified_out}{A string of path to save unclassified sequences,
which should be a fastq file.}

\item{odir}{Output directory.}

\item{envpath}{A character define the environment variables \code{PATH} to be
added before running command.}

\item{envvar}{A named atomic vector define running environment variables of
the command, all environment variables will be replaced before running
command. Use \code{NA} to remove an environment variable.}

\item{abort}{A bool indicates whether to report error if command return
non-zero status.}

\item{stdout, stderr}{Where output to ‘stdout’ or ‘stderr’ should be sent.
Possible values are:
\itemize{
\item \code{TRUE}: print child output in R console
\item \code{FALSE}: suppress output stream
\item \strong{string}: name or path of file to redirect output
}}

\item{stdin}{should input be diverted?  \code{""} means the default,
    alternatively a character string naming a file.  Ignored
    if \code{input} is supplied.}

\item{wait}{A bool (not NA) indicating whether the R interpreter should wait
for the command to finish, or run it asynchronously. This will be ignored
(and the interpreter will always wait) if \code{abort = TRUE}.}

\item{timeout}{timeout in seconds, ignored if 0.  This is a limit for the
    elapsed time running \code{command} in a separate process.   Fractions
    of seconds are ignored.}

\item{verbose}{A logical value indicates whether to print running command
message.}

\item{help}{If \code{TRUE}, will print the help document for this command.}

\item{kraken2}{A string of path to \code{kraken2} command.}

\item{pythonpath}{A character define the environment variables \code{PYTHONPATH}
to be added before running command.}

\item{python}{A string of path to \code{python} command.}
}
\description{
Kraken is a taxonomic sequence classifier that assigns taxonomic labels to
DNA sequences. Kraken examines the k-mers within a query sequence and uses
the information within those k-mers to query a database. That database maps
k-mers to the lowest common ancestor (LCA) of all genomes known to contain a
given k-mer.
}
\seealso{
\url{https://github.com/DerrickWood/kraken2/wiki/Manual}
}
