% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envvars.R
\name{parse_envvar}
\alias{parse_envvar}
\title{Parse an atomic character for system2 env argument}
\usage{
parse_envvar(env)
}
\arguments{
\item{env}{A named atomic character of environment variables.}
}
\value{
A string can be used in \link{system2}
}
\description{
Parse an atomic character for system2 env argument
}
