% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd-perl.R
\name{perl}
\alias{perl}
\title{Perl is a highly capable, feature-rich programming language with over 36
years of development.}
\usage{
perl(..., perl = NULL)
}
\arguments{
\item{...}{<\link[rlang:dyn-dots]{dynamic dots}> Additional arguments passed to \code{perl} command, empty arguments are automatically trimed. Details see: \code{cmd_help(perl())}.}

\item{perl}{A string of path to \code{perl} command.}
}
\value{
A \code{command} object.
}
\description{
Perl is a highly capable, feature-rich programming language with over 36
years of development.
}
\seealso{
\url{https://www.perl.org/}
}
