% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exec-python.R
\name{python}
\alias{python}
\title{Python is a programming language that lets you work quickly and integrate
systems more effectively.}
\usage{
python(
  ...,
  pythonpath = NULL,
  envpath = NULL,
  envvar = NULL,
  help = FALSE,
  stdout = TRUE,
  stderr = TRUE,
  stdin = "",
  wait = TRUE,
  timeout = 0L,
  abort = TRUE,
  verbose = TRUE,
  python = NULL
)
}
\arguments{
\item{...}{Additional arguments passed to \code{python} command. Details see: \code{python(help = TRUE)}}

\item{pythonpath}{A character define the environment variables \code{PYTHONPATH}
to be added before running command.}

\item{envpath}{A character define the environment variables \code{PATH} to be
added before running command.}

\item{envvar}{A named atomic vector define running environment variables of
the command, all environment variables will be replaced before running
command. Use \code{NA} to remove an environment variable.}

\item{help}{If \code{TRUE}, will print the help document for this command.}

\item{stdout, stderr}{Where output to ‘stdout’ or ‘stderr’ should be sent.
Possible values are:
\itemize{
\item \code{TRUE}: print child output in R console
\item \code{FALSE}: suppress output stream
\item \strong{string}: name or path of file to redirect output
}}

\item{stdin}{should input be diverted?  \code{""} means the default,
    alternatively a character string naming a file.  Ignored
    if \code{input} is supplied.}

\item{wait}{A bool (not NA) indicating whether the R interpreter should wait
for the command to finish, or run it asynchronously  Only used when
\code{help=FALSE}. The asynchronous process can be managed with \link{process} tools.}

\item{timeout}{timeout in seconds, ignored if 0.  This is a limit for the
    elapsed time running \code{command} in a separate process.   Fractions
    of seconds are ignored.}

\item{abort}{A bool indicates whether to report error if command return
non-zero status.}

\item{verbose}{A bool indicates whether to print running command message.}

\item{python}{A string of path to \code{python} command.}
}
\description{
Python is a programming language that lets you work quickly and integrate
systems more effectively.
}
\seealso{
\url{https://www.python.org/}
}
