% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd-python.R
\name{python}
\alias{python}
\title{Python is a programming language that lets you work quickly and integrate
systems more effectively.}
\usage{
python(..., python = NULL)
}
\arguments{
\item{...}{<\link[rlang:dyn-dots]{dynamic dots}> Additional arguments passed to \code{python} command, empty arguments are automatically trimed. Details see: \code{cmd_help(python())}.}

\item{python}{A string of path to \code{python} command.}
}
\value{
A \code{command} object.
}
\description{
Python is a programming language that lets you work quickly and integrate
systems more effectively.
}
\seealso{
\url{https://www.python.org/}
}
