% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys_command.R
\name{run_command}
\alias{run_command}
\title{Invoke a System Command}
\usage{
run_command(
  cmd,
  ...,
  envpath = NULL,
  env = NULL,
  sys_args = list(),
  verbose = TRUE
)
}
\arguments{
\item{cmd}{Command to be invoked, as a character string.}

\item{...}{Arguments passed to command.}

\item{envpath}{A character define the environment PATH to add before running
command.}

\item{env}{A named atomic vector define running environment of the command.}

\item{sys_args}{A list of arguments passed to \link{system2}.}

\item{verbose}{A logical value indicates output running command message.}
}
\value{
See \link{system2} for details
}
\description{
Invoke a System Command
}
