% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/command-gistic2.R
\name{run_gistic2}
\alias{run_gistic2}
\title{Run GISTIC2}
\usage{
run_gistic2(
  seg,
  refgene,
  odir = getwd(),
  t_amp = 0.1,
  t_del = 0.1,
  join_segment_size = 4L,
  qv_thresh = 0.25,
  fname = NULL,
  ext = NULL,
  remove_XY = TRUE,
  cap = 1.5,
  run_broad_analysis = FALSE,
  broad_len_cutoff = 0.98,
  twoside = FALSE,
  maxseg = 2500L,
  resolution = 0.05,
  conf_level = 0.75,
  genegistic = FALSE,
  do_arbitration = TRUE,
  peak_type = "robust",
  arm_peeloff = FALSE,
  sample_center = "median",
  savegene = FALSE,
  saveseg = TRUE,
  savedata = TRUE,
  markers = NULL,
  maxspace = 10000L,
  ...,
  gistic2_cmd = NULL,
  gistic2_verbose = 0L,
  envpath = NULL,
  env = NULL,
  abort = FALSE,
  sys_args = list(),
  verbose = TRUE
)
}
\arguments{
\item{seg}{A data.frame of segmented data.}

\item{refgene}{Path to reference genome data input file (REQUIRED, see below
for file description).}

\item{odir}{The output directory.}

\item{t_amp}{Threshold for copy number amplifications. Regions with a copy
number gain above this positive value are considered amplified. Regions with
a copy number gain smaller than this value are considered noise and set to 0.
(DEFAULT=0.1).}

\item{t_del}{Threshold for copy number deletions. Regions with a copy number
loss below the negative of this positive value are considered deletions.
Regions with a smaller copy number loss are considered noise and set to 0.
(DEFAULT=0.1).}

\item{join_segment_size}{Smallest number of markers to allow in segments from
the segmented data. Segments that contain fewer than this number of markers
are joined to the neighboring segment that is closest in copy number.
(DEFAULT=4).}

\item{qv_thresh}{Significance threshold for q-values. Regions with q-values
below this number are considered significant. (DEFAULT=0.25)}

\item{fname}{Base file name to prepend to all output files. (DEFAULT no
output basename).}

\item{ext}{Extension to append to all output files. (DEFAULT='', no
extension).}

\item{remove_XY}{Flag indicating whether to remove data from the sex
chromosomes before analysis. Allowed values= {1,0}. (DEFAULT=1, remove X,Y).}

\item{cap}{Minimum and maximum cap values on analyzed data. Regions with a
log2 ratio greater than the cap are set to the cap value; regions with a log2
ratio less than -cap value are set to -cap. Values must be positive.
(DEFAULT=1.5).}

\item{run_broad_analysis}{Flag indicating that an additional broad-level
analysis should be performed. Allowed values = {1,0}. (DEFAULT = 0, no broad
analysis).}

\item{broad_len_cutoff}{Threshold used to distinguish broad from focal
events, given in units of fraction of chromosome arm. (DEFAULT = 0.98).}

\item{twoside}{Flag indicating that a two-dimensional quadrant figure should
be created as part of a broad analysis. Allowed values = {1,0}.(DEFAULT=0, no
figure).}

\item{maxseg}{Maximum number of segments allowed for a sample in the input
data. Samples with more segments than this threshold are excluded from the
analysis. (DEFAULT=2500).}

\item{resolution}{Resolution used to create the empirical distributions used
to estimate background probabilities. Lower values generate more accurate
results at a cost of greater computation time. (DEFAULT=0.05).}

\item{conf_level}{Confidence level used to calculate the region containing a
driver. (DEFAULT=0.75).}

\item{genegistic}{Flag indicating that the gene GISTIC algorithm should be
used to calculate the significance of deletions at a gene level instead of a
marker level. Allowed values= {1,0}. (DEFAULT=0, no gene GISTIC).}

\item{do_arbitration}{Flag for using the arbitrated peel-off algorithm when
resolving the significance of overlapping peaks. Allowed values= {1,0}.
(DEFAULT=1, use arbitrated peel-off).}

\item{peak_type}{Method for evaluating the significance of peaks, either
robust (DEFAULT) or loo for leave-one-out.}

\item{arm_peeloff}{Flag set to enable arm-level peel-off of events during
peak definition. The arm-level peel-off enhancement to the arbitrated
peel-off method assigns all events in the same chromosome arm of the same
sample to a single peak. It is useful when peaks are split by noise or
chromothripsis. Allowed values= {1,0}. (DEFAULT=0, use normal arbitrated
peel-off).}

\item{sample_center}{Method for centering each sample prior to the GISTIC
analysis. Allowed values are median, mean, or none. (DEFAULT=median).}

\item{savegene}{Flag indicating that gene tables should be saved. Allowed
values= {1,0}. (DEFAULT=0, don't save gene tables).}

\item{saveseg}{Flag indicating that the preprocessed segmented data used as
input for the GISTIC analysis should be saved (in matlab format). Allowed
values= {1,0}. (DEFAULT=1, save segmented input data).}

\item{savedata}{Flag indicating that native MatlabTM output files should be
saved in addition to text data. Allowed values= {1,0}. (DEFAULT=1, save
MatlabTM files).}

\item{markers}{Path to markers input file (OPTIONAL, but encouraged).}

\item{maxspace}{Maximum allowed spacing between pseudo-markers, in bases.
Pseudo-markers are generated when the markers file input is omitted. Segments
that contain fewer than this number of markers are joined to the neighboring
segment that is closest in copy number. (DEFAULT=10,000).}

\item{...}{Other arguments passed to gistic2. Details see: \url{https://broadinstitute.github.io/gistic2/}.}

\item{gistic2_cmd}{Path to gistic2 command.}

\item{gistic2_verbose}{Integer value indicating the level of verbosity to use
in the program execution log. Suggested values = {0,10,20,30}. 0 sets no
verbosity; 30 sets high level of verbosity. (DEFAULT=0)}

\item{envpath}{A character define the environment PATH to add before running
command.}

\item{env}{A named atomic vector define running environment of the command.}

\item{abort}{A scalar logical indicates whether to report error if command
return non-zero status.}

\item{sys_args}{A list of arguments passed to \link{system2}.}

\item{verbose}{A logical value indicates output running command message.}
}
\description{
The GISTIC module identifies regions of the genome that are
significantly amplified or deleted across a set of samples. Each aberration
is assigned a G-score that considers the amplitude of the aberration as well
as the frequency of its occurrence across samples. False Discovery Rate
q-values are then calculated for the aberrant regions, and regions with
q-values below a user-defined threshold are considered significant. For each
significant region, a "peak region" is identified, which is the part of the
aberrant region with greatest amplitude and frequency of alteration. In
addition, a "wide peak" is determined using a leave-one-out algorithm to
allow for errors in the boundaries in a single sample. The "wide peak"
boundaries are more robust for identifying the most likely gene targets in
the region. Each significantly aberrant region is also tested to determine
whether it results primarily from broad events (longer than half a chromosome
arm), focal events, or significant levels of both. The GISTIC module reports
the genomic locations and calculated q-values for the aberrant regions. It
identifies the samples that exhibit each significant amplification or
deletion, and it lists genes found in each "wide peak" region.
}
\seealso{
\url{https://broadinstitute.github.io/gistic2/}
}
