% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exec-trust4.R
\name{trust4}
\alias{trust4}
\alias{trust4_imgt_annot}
\alias{trust4_gene_names}
\title{TRUST4: immune repertoire reconstruction from bulk and single-cell RNA-seq
data}
\usage{
trust4(
  file1,
  ref_coordinate,
  ...,
  file2 = NULL,
  mode = NULL,
  ref_annot = NULL,
  ofile = NULL,
  odir = getwd(),
  envpath = NULL,
  envvar = NULL,
  abort = TRUE,
  stdout = TRUE,
  stderr = TRUE,
  stdin = "",
  wait = TRUE,
  timeout = 0L,
  verbose = TRUE,
  help = FALSE,
  trust4 = NULL
)

trust4_imgt_annot(
  species = "Homo_sapien",
  ofile = "IMGT+C.fa",
  odir = getwd(),
  envpath = NULL,
  envvar = NULL,
  abort = TRUE,
  stdout = TRUE,
  stderr = TRUE,
  stdin = "",
  wait = TRUE,
  timeout = 0L,
  verbose = TRUE,
  perl = NULL
)

trust4_gene_names(imgt_annot, ofile = "bcr_tcr_gene_name.txt", odir = getwd())
}
\arguments{
\item{file1}{Path to bam file or fastq file.}

\item{ref_coordinate}{Path to the fasta file coordinate and sequence of
V/D/J/C genes.}

\item{...}{Additional arguments passed to \code{run-trust4} command. Details see:
\url{https://github.com/liulab-dfci/TRUST4}.}

\item{file2}{Path to the second paired-end read fastq file, only used for
\code{mode = "fastq"}.}

\item{mode}{One of "bam" or "fastq". If \code{NULL}, will be inferred from
\code{file1}.}

\item{ref_annot}{Path to detailed V/D/J/C gene reference file, such as from
IMGT database. (default: not used). (recommended).}

\item{ofile}{\itemize{
\item \code{trust4}: Prefix of output files. (default: inferred from file
prefix).
\item \code{trust4_imgt_annot}: Output file name.
\item \code{trust4_gene_names}: Output file name.
}}

\item{odir}{A string of path to the output directory.}

\item{envpath}{A character define the environment variables \code{PATH} to be
added before running command.}

\item{envvar}{A named atomic vector define running environment variables of
the command, all environment variables will be replaced before running
command. Use \code{NA} to remove an environment variable.}

\item{abort}{A bool indicates whether to report error if command return
non-zero status.}

\item{stdout, stderr}{Where output to ‘stdout’ or ‘stderr’ should be sent.
Possible values are:
\itemize{
\item \code{TRUE}: print child output in R console
\item \code{FALSE}: suppress output stream
\item \strong{string}: name or path of file to redirect output
}}

\item{stdin}{should input be diverted?  \code{""} means the default,
    alternatively a character string naming a file.  Ignored
    if \code{input} is supplied.}

\item{wait}{A bool (not NA) indicating whether the R interpreter should wait
for the command to finish, or run it asynchronously. This will be ignored
(and the interpreter will always wait) if \code{abort = TRUE}.}

\item{timeout}{timeout in seconds, ignored if 0.  This is a limit for the
    elapsed time running \code{command} in a separate process.   Fractions
    of seconds are ignored.}

\item{verbose}{A logical value indicates whether to print running command
message.}

\item{help}{If \code{TRUE}, will print the help document for this command.}

\item{trust4}{A string of path to \code{run-trust4} command.}

\item{species}{Species to extract IMGT annotation, details see
\url{https://www.imgt.org//download/V-QUEST/IMGT_V-QUEST_reference_directory/}.}

\item{perl}{Path to \code{perl} command.}

\item{imgt_annot}{Path of IMGT annotation file, created via
\code{trust4_imgt_annot}.}
}
\description{
TRUST4: immune repertoire reconstruction from bulk and single-cell RNA-seq
data
}
\seealso{
\url{https://github.com/liulab-dfci/TRUST4}
}
